WIP added persistent UI state and default configurations
