+++ frontend/package.json
@@ -16,13 +16,16 @@
         "@orama/orama": "^3.1.11",
         "@orama/plugin-match-highlight": "^3.1.11",
         "@tailwindcss/vite": "^4.1.11",
+        "@tanstack/db": "^0.4.1",
+        "@tanstack/query-db-collection": "^0.2.22",
         "@tanstack/react-query": "^5.87.0",
         "@tanstack/react-router": "^1.121.12",
         "@tanstack/react-router-devtools": "^1.121.12",
         "dotenv": "^16.3.1",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
-        "typescript-parsec": "^0.3.4"
+        "typescript-parsec": "^0.3.4",
+        "zod": "^4.1.11"
     },
     "devDependencies": {
         "@types/node": "^20.10.6",
